checkAnswersService test split into different files
