+++ src/test/unit/services/features/response/checkAnswers/constants.ts
@@ -1,0 +1,7 @@
+export const CLAIM_ID = 'claimId';
+
+// -- Summary Section Array Index
+export const INDEX_DETAILS_SECTION = 0;
+export const INDEX_RESPONSE_CLAIM_SECTION = 1;
+export const INDEX_FINANCIAL_SECTION = 2;
+export const INDEX_RESPONSE_SECTION = 3;
